Allow setting album manually
If the user provides the --album flag, all songs will be tagged with the
given name. Album will still be parsed from the other sources, but added
to the discsubtitle tag instead. This is to abuse the tag somewhat, but
music players will generally show the tags in a nice way, as "album -
discsubtitle". This way, songs from different albums can be grouped
together kind of like in a playlist.