Preserve original album data for disc subtitle
Earlier behavior was to discard the original album tag if the user
provides a manual album tag that should be used for all songs. In this
scenario, the album parsed from other sources is put in the disc
subtitle tag instead, but the original album tag was not discarded and
not given as a suggested tag. This is fixed now.