Don't skip song not containing auto-added data
Some data is added automatically and is not contained within the four
existing tag sources, such as the comment tag which is automatically
added when processing songs and set to "youtube-dl". This potential
difference between new and old metadata is now accounted for when
deciding whether to skip a song because there is no data to add.