Add test and coverage to CI pipeline
Pytest is run with the --cov option in a job called "test" in the Gitlab
workflow of the project. The job fails when coverage is less than 7%,
which is the current coverage. This is to prevent regression. The
percentage will be increased bit by bit as more tests are added.