Test automatically resolving the artist
When it is obvious who the artist is, that is, when it's the same artist
in both the youtube description and the original tags, it's okay to
automatically put that in the resolved tags.

The tested code also had a bug fixed where identicalness wasn't detected
if there were multiple artists in a list in the original tags.