Test that other action reshows action menu
When the user choses to perform an "other action", such as looking at
the text of the youtube description, they should after that again be
presented with the chance to choose which of the sources they want to
resolve a given tag for.

Also update the minimum coverage so that lower than the current one is
not allowed.