Actually delete tags marked for deletion
There was a bug where tags marged for deletion weren't actually deleted,
since they were removed from the "resolved" keys, instead of marked for
deletion. "resolved" is different from the actual metadata object that
will be deleted, a mark needs to be left so the actual metadata object
can be pruned later. At least according to the current structure.

Also add tests to ensure that tags actually get deleted and that they
are printed properly when marked for deletion.