Print tags that are not tracked as well
When printing all tags, show ones that are not parsed by Retag-Opus as
well. It is still possible to delete them from the app, or to modify
them. Then the user will want to see them. They may also want to see if
there is anything they would like to delete in that set.